move object and block store interfaces to their own files
