+++ pkg/cloudprovider/object_store.go
@@ -19,8 +19,6 @@ package cloudprovider
 import (
 	"io"
 	"time"
-
-	"k8s.io/apimachinery/pkg/runtime"
 )
 
 // ObjectStore exposes basic object-storage operations required
@@ -55,35 +53,4 @@ type ObjectStore interface {
 
 	// CreateSignedURL creates a pre-signed URL for the given bucket and key that expires after ttl.
 	CreateSignedURL(bucket, key string, ttl time.Duration) (string, error)
-}
-
-// BlockStore exposes basic block-storage operations required
-// by Ark.
-type BlockStore interface {
-	// Init prepares the BlockStore for usage using the provided map of
-	// configuration key-value pairs. It returns an error if the BlockStore
-	// cannot be initialized from the provided config.
-	Init(config map[string]string) error
-
-	// CreateVolumeFromSnapshot creates a new block volume in the specified
-	// availability zone, initialized from the provided snapshot,
-	// and with the specified type and IOPS (if using provisioned IOPS).
-	CreateVolumeFromSnapshot(snapshotID, volumeType, volumeAZ string, iops *int64) (volumeID string, err error)
-
-	// GetVolumeID returns the cloud provider specific identifier for the PersistentVolume.
-	GetVolumeID(pv runtime.Unstructured) (string, error)
-
-	// SetVolumeID sets the cloud provider specific identifier for the PersistentVolume.
-	SetVolumeID(pv runtime.Unstructured, volumeID string) (runtime.Unstructured, error)
-
-	// GetVolumeInfo returns the type and IOPS (if using provisioned IOPS) for
-	// the specified block volume in the given availability zone.
-	GetVolumeInfo(volumeID, volumeAZ string) (string, *int64, error)
-
-	// CreateSnapshot creates a snapshot of the specified block volume, and applies the provided
-	// set of tags to the snapshot.
-	CreateSnapshot(volumeID, volumeAZ string, tags map[string]string) (snapshotID string, err error)
-
-	// DeleteSnapshot deletes the specified volume snapshot.
-	DeleteSnapshot(snapshotID string) error
 }
